test di tutte le sorgenti

+++ src/metrics.h
@@ -11,7 +11,7 @@
 // Struttura per una singola metrica
 typedef struct {
     char name[64];
-    int value;
+    double value;
     char unit[16];  // Unità di misura (%, MB, GB, KB/s, ecc.)
 } Metric;
 
@@ -49,7 +49,7 @@ void metrics_stop_collection(void);
 
 void metrics_updated_callback(const Metrics* metrics);
 void metrics_set(const char* name, int value);
-void metrics_set_with_unit(const char* name, int value, const char* unit);
+void metrics_set_with_unit(const char* name, double value, const char* unit);
 
 // Token metrics functions
 void generate_random_token(char* token, size_t length);
